refactor: major refactor part 1
