+++ src/components/App.module.scss
@@ -3,8 +3,8 @@
   height: 40px;
   mask-size: cover;
   mask-position: center;
-  -webkit-mask-image: url(./car.svg);
-  mask-image: url(./car.svg);
+  -webkit-mask-image: url(../img/car.svg);
+  mask-image: url(../img/car.svg);
 }
 
 @keyframes animateCar {
