+++ src/components/Winners/WinnersTable.module.scss
@@ -4,6 +4,6 @@
   height: 20px;
   mask-size: cover;
   mask-position: center;
-  -webkit-mask-image: url(../car.svg);
-  mask-image: url(../car.svg);
+  -webkit-mask-image: url(../../img/car.svg);
+  mask-image: url(../../img/car.svg);
 }
