+++ src/constances.ts
@@ -2,4 +2,10 @@ export const BASE_URL = 'http://127.0.0.1:3000';
 
 export const CARS_PER_PAGE = 7;
 
-export const WINNERS_PER_PAGE = 10;
+export const WINNERS_PER_PAGE = 3;
+
+export const CAR_MODELS = [
+  ['Tesla', 'BMW', 'Mercedes', 'Opel', 'Skoda', 'Audi', 'Crysler', 'Dodge', 'Ford', 'Ferrari',
+    'Lamborghini', 'Peugeout', 'Fisker', 'Aston Martin', 'Lada'],
+  ['Model S', 'Model E', 'Model X', 'Model Y', '230i', '540d', 'Corsa', 'Octavia', 'Fabia', 'Rapid',
+    '300c', 'Viper', 'Mustang', 'Karma', 'Aventador']];
