--- src/types.ts
@@ -1,11 +1,0 @@
-export type TWinner = {
-  id: number;
-  wins: number;
-  time: number;
-}
-
-export type TCar = {
-  name: string;
-  color: string;
-  id: number;
-}
